TcxFileSegregation
The script reads through a folder containing all the .tcx files downloaded from the Strava server. These files can be opened and read as XML files and the lxml package's Element Tree API (lxml.etree) is used to parse the files iteratively to identify the 'Activity' element and its attribute which is the type of activity. All the 'Ride' activities are identified and these files are copied to anewly created folder for further processing in the chain.

In some instances, the .XML files contain unwanted characters/whitespace preceding the xml declaration. This causes an error. To resolve this, the 'clean_xml_content' function is written which removes the unwanted characters/whitespace.